refactor(table): large update
