+++ packages/renderer/src/lib/objects/KubernetesObjectUI.ts
@@ -24,5 +24,4 @@ export interface KubernetesObjectUI {
 
 export interface KubernetesNamespacedObjectUI extends KubernetesObjectUI {
   namespace: string;
-  selected: boolean;
 }
