+++ packages/renderer/src/lib/volume/VolumeInfoUI.ts
@@ -28,7 +28,6 @@ export interface VolumeInfoUI {
   humanSize: string;
   engineId: string;
   engineName: string;
-  selected: boolean;
   status: 'USED' | 'UNUSED' | 'DELETING';
   containersUsage: { id: string; names: string[] }[];
 }
