+++ packages/renderer/src/stores/tasks.ts
@@ -25,12 +25,9 @@ import type { TaskImpl } from '../../../main/src/plugin/tasks/task-impl';
 import { findMatchInLeaves } from './search-util';
 
 /**
- * When tasks are displayed into a table, the selected property is used
- * to flag the task as being selectable.
+ * Object for Table
  */
-export interface TaskInfoUI extends TaskInfo {
-  selected?: boolean;
-}
+export type TaskInfoUI = TaskInfo;
 
 /**
  * Defines the store used to define the tasks.
