+++ packages/ui/src/lib/table/Table.svelte
@@ -1,85 +1,149 @@
-<!--
-    @component
-    @deprecated
--->
 <style>
 .grid-table {
   display: grid;
+  grid-template-columns: var(--table-grid-table-columns);
 }
 </style>
-
-<script lang="ts">
-/* eslint-disable import/no-duplicates */
-// https://github.com/import-js/eslint-plugin-import/issues/1479
-import { faChevronDown, faChevronRight } from '@fortawesome/free-solid-svg-icons';
-import { afterUpdate, onMount, tick } from 'svelte';
+<script lang="ts" generics="T">
+import { faChevronDown } from '@fortawesome/free-solid-svg-icons/faChevronDown';
+import { faChevronRight } from '@fortawesome/free-solid-svg-icons/faChevronRight';
+import { onMount } from 'svelte';
+import { SvelteSet } from 'svelte/reactivity';
 import Fa from 'svelte-fa';
 
 import Checkbox from '../checkbox/Checkbox.svelte';
-/* eslint-enable import/no-duplicates */
 import type { Column, Row } from './table';
 
-export let kind: string;
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export let columns: Column<any>[];
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export let row: Row<any>;
-export let data: { selected?: boolean; name?: string }[];
-export let defaultSortColumn: string | undefined = undefined;
-export let collapsed: string[] = [];
+interface Props<T> {
+  readonly kind: string;
+  readonly data: Array<T>;
+  collapsed?: Set<string>;
+  selected?: Set<string>;
+  readonly defaultSortColumn?: string;
+  readonly row: Row<T>;
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  readonly columns: Array<Column<T, any>>;
 
-let tableHtmlDivElement: HTMLDivElement | undefined = undefined;
+  /**
+   * @deprecated useful for parent component using Svelte 4
+   */
+  selectedItemsNumber?: number;
 
-// number of selected items in the list
-export let selectedItemsNumber: number = 0;
-$: selectedItemsNumber = row.info.selectable
-  ? data.filter(object => row.info.selectable?.(object) && object.selected).length +
-    data.reduce(
-      (previous, current) =>
-        previous +
-        (row.info.children?.(current)?.filter(child => row.info.selectable?.(child) && child.selected).length ?? 0),
-      0,
+  /**
+   * the key function is mandatory to determine the key for selected and collapsed
+   *
+   * @remarks the key must be unique for every object T
+   * @param object
+   */
+  key(object: T): string;
+
+  /**
+   * To set the aria-label on each row, you can specify a label function
+   * @param object
+   */
+  label?(object: T): string;
+}
+
+let {
+  kind,
+  data,
+  collapsed = $bindable(new SvelteSet()),
+  selected = $bindable(new SvelteSet()),
+  row,
+  columns,
+  key,
+  // default label to key
+  label = key,
+  defaultSortColumn,
+  selectedItemsNumber = $bindable(),
+}: Props<T> = $props();
+
+let selectedCount: number = $derived(selected.size);
+
+$effect(() => {
+  selectedItemsNumber = selectedCount;
+});
+
+let selectedAllCheckboxes: boolean = $derived.by(() => {
+  // shortcut if row has no selectable function
+  if (!row.info.selectable) return false;
+
+  // shortcut if none is selectable
+  if (data.filter(object => row.info.selectable?.(object)).length === 0) return false;
+
+  return data
+    .reduce(
+      (accumulator, current) => {
+        const children = row.info.children?.(current);
+        if (children) {
+          accumulator.push(...children);
+        }
+        return accumulator;
+      },
+      [...data] as Array<T>,
     )
-  : 0;
-
-// do we need to unselect all checkboxes if we don't have all items being selected ?
-let selectedAllCheckboxes = false;
-$: selectedAllCheckboxes = row.info.selectable
-  ? data.filter(object => row.info.selectable?.(object)).every(object => object.selected) &&
-    data
-      .filter(object => row.info.children?.(object))
-      .map(object => row.info.children?.(object))
-      .flat()
-      .filter(child => row.info.selectable?.(child))
-      .every(child => child.selected) &&
-    data.filter(object => row.info.selectable?.(object)).length > 0
-  : false;
+    .filter(child => row.info.selectable?.(child))
+    .every(child => selected.has(key(child)));
+});
 
 function toggleAll(e: CustomEvent<boolean>): void {
-  const checked = e.detail;
+  // shortcut if not selectable
   if (!row.info.selectable) {
     return;
   }
-  data.filter(object => row.info.selectable?.(object)).forEach(object => (object.selected = checked));
 
-  // toggle children
-  data.forEach(object => {
-    const children = row.info.children?.(object);
-    if (children) {
-      children.filter(child => row.info.selectable?.(child)).forEach(child => (child.selected = checked));
+  data.forEach(object => objectChecked(object, e.detail));
+}
+
+function objectChecked(object: T, checked: boolean): void {
+  // select object itself
+  if (row.info.selectable?.(object)) {
+    if (checked) {
+      selected.add(key(object));
+    } else {
+      selected.delete(key(object));
     }
-  });
-  data = [...data];
+  }
+
+  // select object children
+  if (row.info.children?.(object)) {
+    row.info
+      .children?.(object)
+      ?.filter(child => row.info.selectable?.(child))
+      .forEach(child => {
+        if (checked) {
+          selected.add(key(child));
+        } else {
+          selected.delete(key(child));
+        }
+      });
+  }
 }
 
-let sortCol: Column<unknown>;
-let sortAscending: boolean;
+// sorting states
+let sortCol: Column<T> | undefined = $state(undefined);
+let sortAscending: boolean = $state(false);
 
-$: if (data && sortCol) {
-  sortImpl();
-}
+let sortedData: Array<T> = $derived.by(() => {
+  if (!sortCol) return data;
 
-function sort(column: Column<unknown>): void {
+  let comparator = sortCol.info.comparator;
+  if (!comparator) {
+    // column is not sortable
+    return data;
+  }
+
+  if (!sortAscending) {
+    // we're already sorted, switch to reverse order
+    let comparatorTemp = comparator;
+    comparator = (a, b): number => -comparatorTemp(a, b);
+  }
+
+  console.log('sortData', data);
+  return data.toSorted(comparator);
+});
+
+function sort(column: Column<T>): void {
   if (!column) {
     return;
   }
@@ -96,48 +160,22 @@ function sort(column: Column<unknown>): void {
     sortCol = column;
     sortAscending = column.info.initialOrder ? column.info.initialOrder !== 'descending' : true;
   }
-  sortImpl();
 }
 
-function sortImpl(): void {
-  // confirm we're sorting
-  if (!sortCol) {
-    return;
+/**
+ * Update the collapse state of a row
+ * @param key
+ */
+function toggleChildren(key: string): void {
+  if (collapsed.has(key)) {
+    collapsed.delete(key);
+  } else {
+    collapsed.add(key);
   }
-
-  let comparator = sortCol.info.comparator;
-  if (!comparator) {
-    // column is not sortable
-    return;
-  }
-
-  if (!sortAscending) {
-    // we're already sorted, switch to reverse order
-    let comparatorTemp = comparator;
-    comparator = (a, b): number => -comparatorTemp(a, b);
-  }
-
-  // eslint-disable-next-line etc/no-assign-mutated-array
-  data = data.sort(comparator);
 }
 
-onMount(async () => {
-  const column: Column<unknown> | undefined = columns.find(column => column.title === defaultSortColumn);
-  if (column?.info.comparator) {
-    sortCol = column;
-    sortAscending = column.info.initialOrder ? column.info.initialOrder !== 'descending' : true;
-    await tick(); // Ensure all initializations are complete.
-    sortImpl(); // Explicitly call sortImpl to sort the data initially.
-  }
-});
-
-afterUpdate(() => {
-  tick()
-    .then(() => setGridColumns())
-    .catch((err: unknown) => console.error('unable to refresh grid columns', err));
-});
-
-function setGridColumns(): void {
+// grid computation
+let gridTemplateColumns: string = $derived.by(() => {
   // section and checkbox columns
   let columnWidths: string[] = ['20px'];
 
@@ -151,45 +189,20 @@ function setGridColumns(): void {
   // final spacer
   columnWidths.push('5px');
 
-  let wid = columnWidths.join(' ');
-  if (tableHtmlDivElement) {
-    let grids: HTMLCollection = tableHtmlDivElement.getElementsByClassName('grid-table');
-    for (const element of grids) {
-      (element as HTMLElement).style.setProperty('grid-template-columns', wid);
-    }
-  }
-}
+  return columnWidths.join(' ');
+});
 
-function objectChecked(object: { selected?: boolean }): void {
-  // check for children and set them to the same state
-  if (row.info.children) {
-    const children = row.info.children(object);
-    if (children) {
-      // event fires before parent changes so use '!'
-      children.forEach(child => (child.selected = !object.selected));
-    }
+onMount(() => {
+  const column: Column<T> | undefined = columns.find(column => column.title === defaultSortColumn);
+  if (column?.info.comparator) {
+    sortCol = column;
+    sortAscending = column.info.initialOrder ? column.info.initialOrder !== 'descending' : true;
   }
-}
-
-function toggleChildren(name: string | undefined): void {
-  if (!name) {
-    return;
-  }
-
-  if (collapsed.includes(name)) {
-    const index = collapsed.indexOf(name, 0);
-    if (index > -1) {
-      collapsed.splice(index, 1);
-    }
-  } else {
-    collapsed.push(name);
-  }
-  // trigger Svelte update
-  collapsed = collapsed;
-}
+});
 </script>
-
-<div class="w-full mx-5" class:hidden={data.length === 0} role="table" aria-label={kind} bind:this={tableHtmlDivElement}>
+<div
+  style="--table-grid-table-columns: {gridTemplateColumns}"
+  class="w-full mx-5" class:hidden={data.length === 0} role="table" aria-label={kind}>
   <!-- Table header -->
   <div role="rowgroup">
     <div
@@ -202,13 +215,13 @@ function toggleChildren(name: string | undefined): void {
             title="Toggle all"
             bind:checked={selectedAllCheckboxes}
             disabled={!row.info.selectable || data.filter(object => row.info.selectable?.(object)).length === 0}
-            indeterminate={selectedItemsNumber > 0 && !selectedAllCheckboxes}
+            indeterminate={selectedCount > 0 && !selectedAllCheckboxes}
             on:click={toggleAll} />
         </div>
       {/if}
       {#each columns as column, index (index)}
-        <!-- svelte-ignore a11y-click-events-have-key-events -->
-        <!-- svelte-ignore a11y-interactive-supports-focus -->
+        <!-- svelte-ignore a11y_click_events_have_key_events -->
+        <!-- svelte-ignore a11y_interactive_supports_focus -->
         <div
           class="max-w-full overflow-hidden flex flex-row text-sm font-semibold items-center whitespace-nowrap {column
             .info.align === 'right'
@@ -217,45 +230,48 @@ function toggleChildren(name: string | undefined): void {
               ? 'justify-self-center'
               : 'justify-self-start'} self-center select-none"
           class:cursor-pointer={column.info.comparator}
-          on:click={sort.bind(undefined, column)}
+          onclick={sort.bind(undefined, column)}
           role="columnheader">
           <div class="overflow-hidden text-ellipsis">
             {column.title}
           </div>
           {#if column.info.comparator}<i
-              class="fas pl-0.5"
-              class:fa-sort={sortCol !== column}
-              class:fa-sort-up={sortCol === column && sortAscending}
-              class:fa-sort-down={sortCol === column && !sortAscending}
-              class:text-[var(--pd-table-header-unsorted)]={sortCol !== column}
-              aria-hidden="true"></i
-            >{/if}
+            class="fas pl-0.5"
+            class:fa-sort={sortCol !== column}
+            class:fa-sort-up={sortCol === column && sortAscending}
+            class:fa-sort-down={sortCol === column && !sortAscending}
+            class:text-[var(--pd-table-header-unsorted)]={sortCol !== column}
+            aria-hidden="true"></i
+          >{/if}
         </div>
       {/each}
     </div>
   </div>
+
   <!-- Table body -->
   <div role="rowgroup">
-    {#each data as object (object)}
+    {#each sortedData as object (key(object))}
       {@const children = row.info.children?.(object) ?? []}
       <div class="min-h-[48px] h-fit bg-[var(--pd-content-card-bg)] rounded-lg mb-2 border border-[var(--pd-content-table-border)]">
+
+        <!-- row parent -->
         <div
           class="grid grid-table gap-x-0.5 min-h-[48px] hover:bg-[var(--pd-content-card-hover-bg)]"
-          class:rounded-t-lg={object.name &&
-            !collapsed.includes(object.name) &&
+          class:rounded-t-lg={
+            !collapsed.has(key(object)) &&
             children.length > 0}
-          class:rounded-lg={!object.name ||
-            collapsed.includes(object.name) ||
+          class:rounded-lg={
+            collapsed.has(key(object)) ||
             children.length === 0}
           role="row"
-          aria-label={object.name}>
+          aria-label={label(object)}>
           <div class="whitespace-nowrap place-self-center" role="cell">
             {#if children.length > 0}
-              <button on:click={toggleChildren.bind(undefined, object.name)}>
+              <button onclick={toggleChildren.bind(undefined, key(object))}>
                 <Fa
                   size="0.8x"
                   class="text-[var(--pd-table-body-text)] cursor-pointer"
-                  icon={object.name && !collapsed.includes(object.name) ? faChevronDown : faChevronRight} />
+                  icon={!collapsed.has(key(object)) ? faChevronDown : faChevronRight} />
               </button>
             {/if}
           </div>
@@ -263,10 +279,10 @@ function toggleChildren(name: string | undefined): void {
             <div class="whitespace-nowrap place-self-center" role="cell">
               <Checkbox
                 title="Toggle {kind}"
-                bind:checked={object.selected}
+                checked={selected.has(key(object))}
                 disabled={!row.info.selectable(object)}
                 disabledTooltip={row.info.disabledText}
-                on:click={objectChecked.bind(undefined, object)} />
+                onclick={objectChecked.bind(undefined, object)} />
             </div>
           {/if}
           {#each columns as column, index (index)}
@@ -280,30 +296,31 @@ function toggleChildren(name: string | undefined): void {
                 : 'overflow-hidden'} max-w-full py-1.5"
               role="cell">
               {#if column.info.renderer}
-                <svelte:component
-                  this={column.info.renderer}
-                  object={column.info.renderMapping ? column.info.renderMapping(object) : object} />
+                {@const Renderer = column.info.renderer}
+                <Renderer object={column.info.renderMapping ? column.info.renderMapping(object) : object}/>
               {/if}
             </div>
           {/each}
         </div>
 
         <!-- Child objects -->
-        {#if object.name && !collapsed.includes(object.name) && children.length > 0}
-          {#each children as child, i (child)}
+        {#if !collapsed.has(key(object)) && children.length > 0}
+          {#each children as child, i (key(child))}
             <div
               class="grid grid-table gap-x-0.5 hover:bg-[var(--pd-content-card-hover-bg)]"
               class:rounded-b-lg={i === children.length - 1}
               role="row"
-              aria-label={child.name}>
+              aria-label={label(child)}>
               <div class="whitespace-nowrap justify-self-start" role="cell"></div>
               {#if row.info.selectable}
                 <div class="whitespace-nowrap place-self-center" role="cell">
                   <Checkbox
                     title="Toggle {kind}"
-                    bind:checked={child.selected}
+                    checked={selected.has(key(child))}
                     disabled={!row.info.selectable(child)}
-                    disabledTooltip={row.info.disabledText} />
+                    disabledTooltip={row.info.disabledText}
+                    onclick={objectChecked.bind(undefined, child)}
+                  />
                 </div>
               {/if}
               {#each columns as column, index (index)}
@@ -317,9 +334,8 @@ function toggleChildren(name: string | undefined): void {
                     : 'overflow-hidden'} max-w-full py-1.5"
                   role="cell">
                   {#if column.info.renderer}
-                    <svelte:component
-                      this={column.info.renderer}
-                      object={column.info.renderMapping ? column.info.renderMapping(child) : child} />
+                    {@const Renderer = column.info.renderer}
+                    <Renderer object={column.info.renderMapping ? column.info.renderMapping(child) : child}/>
                   {/if}
                 </div>
               {/each}
